+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/ui/instance/toolwindow/MainToolWindowFactory.kt
@@ -18,7 +18,6 @@ import com.intellij.openapi.wm.impl.content.ToolWindowContentUi
 import com.intellij.ui.content.ContentFactory
 import com.intellij.ui.dsl.builder.RightGap
 import com.intellij.ui.dsl.builder.Row
-import com.intellij.ui.dsl.builder.actionButton
 import com.intellij.util.ui.components.BorderLayoutPanel
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.settings.DeveloperToolsApplicationSettings
 import dev.turingcomplete.intellijdevelopertoolsplugin._internal.settings.DeveloperToolsToolWindowSettings
@@ -43,7 +42,6 @@ internal class MainToolWindowFactory : ToolWindowFactory, DumbAware {
     toolWindow.stripeTitle = "Developer Tools"
   }
 
-  @Suppress("UnstableApiUsage")
   override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
     val loaderPanel = BorderLayoutPanel().apply {
       addToCenter(JLabel(CollaborationToolsUIUtil.animatedLoadingIcon))